+++ tests/compile_errors/invalid_policy.cpp
@@ -1,0 +1,8 @@
+#include <easy_uds/easy_uds.hpp>
+
+int main() {
+    auto handler = [](const easy_uds::Request&) {
+        return easy_uds::Response::ok();
+    };
+    (void)easy_uds::RouteOptions{handler}.serialize_in("drive", "latest_wins");
+}